Loop variable is unsigned

+++ tiny_dnn/core/kernels/global_avepool_op_internal.h
@@ -16,7 +16,7 @@ inline void global_avepool_op_internal(
   Tensor<float_t, S2> &out_data,
   const core::global_avepool_params &params,
   const bool layer_parallelize) {
-  for_i(layer_parallelize, in_data.shape()[0], [&](int sample) {
+  for_i(layer_parallelize, in_data.shape()[0], [&](size_t sample) {
 
     const size_t pool_area = params.in.width_ * params.in.height_;
     for (size_t i = 0; i < params.in.depth_; i++) {
@@ -35,7 +35,7 @@ inline void global_avepool_grad_op_internal(
   const Tensor<float_t, S2> &curr_delta,
   const core::global_avepool_params &params,
   const bool layer_parallelize) {
-  for_i(layer_parallelize, prev_delta.shape()[0], [&](int sample) {
+  for_i(layer_parallelize, prev_delta.shape()[0], [&](size_t sample) {
 
     const size_t pool_area = params.in.width_ * params.in.height_;
     for (size_t i = 0; i < params.in.depth_; i++) {
